+++ js/src/Components/ParagraphForm.js
@@ -17,20 +17,24 @@ export const ParagraphForm = ({item, widgetContext, open, typeLabel, onClose}) =
       open={open}
       onClose={() => onClose(false)}
     >
-      <AdminTitleField
-        textField={open}
-        item={item}
-        onChange={widgetContext.onAdminTitleChange.bind(undefined, item.id)}
-      />
-
       <div className="item-form">
         <EntityForm
           entityType="paragraph"
           bundle={item.entity.type[0].target_id}
           entity={item.entity}
+          itemId={item.id}
           onFieldChange={(fieldName, newValue) => widgetContext.updateRowItemEntity(item, fieldName, newValue)}
+          onBehaviorChange={widgetContext.updateEntityBehaviors.bind(undefined, item)}
           widgetContext={widgetContext}
-        />
+          header={
+            <AdminTitleField
+              textField={open}
+              item={item}
+              onChange={widgetContext.onAdminTitleChange.bind(undefined, item.id)}
+            />
+          }
+        >
+        </EntityForm>
       </div>
     </FormDialog>
   )