+++ js/src/Components/Widgets/CheckboxesWidget.js
@@ -33,6 +33,8 @@ export const CheckboxesWidget = ({fieldId, defaultValue, onFieldChange, settings
         {Object.keys(settings.options).map(key =>
           <FormControlLabel
             key={`${fieldId}-${key}`}
+            labelPlacement="end"
+            label={settings.options[key]}
             control={
               <Checkbox
                 id={`${fieldId}-${key}`}
@@ -41,11 +43,10 @@ export const CheckboxesWidget = ({fieldId, defaultValue, onFieldChange, settings
                 onChange={checkboxChanged}
               />
             }
-            label={settings.options[key]}
           />
         )}
       </FormGroup>
-      {settings.help && <FormHelperText>{settings.help}</FormHelperText>}
+      {settings.help && <FormHelperText dangerouslySetInnerHTML={{__html: settings.help}}/>}
     </FormControl>
   )
 };