+++ js/src/Components/Widgets/SelectWidget.js
@@ -1,15 +1,20 @@
 import React from 'react';
 import InputLabel from '@material-ui/core/InputLabel';
-import FormGroup from '@material-ui/core/FormGroup';
+import makeStyles from "@material-ui/core/styles/makeStyles";
 import FormHelperText from '@material-ui/core/FormHelperText';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
+import FormControl from "@material-ui/core/FormControl";
 
 export const SelectWidget = ({fieldId, defaultValue, onFieldChange, settings}) => {
 
   let defaultFieldValue = settings.default_value;
   if (defaultValue && defaultValue.length >= 1) {
     defaultFieldValue = settings.cardinality === 1 ? defaultValue[0][settings.column_key] : defaultValue.map(item => item[settings.column_key]);
+  }
+
+  if (defaultFieldValue === undefined && settings.cardinality !== 1) {
+    defaultFieldValue = [];
   }
 
   const valueChanged = (newValue) => {
@@ -31,18 +36,19 @@ export const SelectWidget = ({fieldId, defaultValue, onFieldChange, settings}) =
 
     onFieldChange(newFieldValue);
   };
+  const classes = useStyles();
 
   return (
-    <FormGroup>
-      <InputLabel htmlFor={fieldId}>{settings.label}</InputLabel>
+    <FormControl required={settings.required} variant="outlined"  className={classes.formControl}>
+      <InputLabel id={'label-'+ fieldId} htmlFor={fieldId}>{settings.label}</InputLabel>
       <Select
         id={fieldId}
+        labelId={'label-'+ fieldId}
         value={defaultFieldValue}
         multiple={settings.cardinality !== 1}
         onChange={e => valueChanged(e.target.value)}
-        variant="outlined"
-        required={settings.required}
         style={{maxWidth: "400px", marginTop: "10px"}}
+        inputProps={{required: settings.required}}
       >
         {settings.required === false && settings.cardinality === 1 &&
         <MenuItem value="">
@@ -58,10 +64,14 @@ export const SelectWidget = ({fieldId, defaultValue, onFieldChange, settings}) =
 
       </Select>
       {settings.help.length > 0 &&
-      <FormHelperText>
-        {settings.help}
-      </FormHelperText>
+      <FormHelperText dangerouslySetInnerHTML={{__html: settings.help}}/>
       }
-    </FormGroup>
+    </FormControl>
   )
 };
+
+const useStyles = makeStyles((theme) => ({
+  formControl: {
+    width: '100%'
+  },
+}));