+++ js/src/Components/Widgets/ViewFieldWidget.js
@@ -63,9 +63,7 @@ export class ViewFieldWidget extends Component {
           {this.props.settings.label}
         </InputLabel>
 
-        <FormHelperText>
-          {this.props.settings.help}
-        </FormHelperText>
+        <FormHelperText dangerouslySetInnerHTML={{__html: this.props.settings.help}}/>
 
         <Select
           id={this.props.fieldId + '-target-id'}
@@ -89,7 +87,7 @@ export class ViewFieldWidget extends Component {
           )}
         </Select>
 
-        <InputWrapper
+        <InputGroupWrapper
           style={{display: this.state.target_id !== '_none' ? 'block' : 'none'}}>
 
           <InputLabel htmlFor={this.props.fieldId + '-display-id'}>
@@ -121,8 +119,13 @@ export class ViewFieldWidget extends Component {
 
           <AdvancedOptions>
             <a href="#"
-               onClick={() => this.setState({advancedOpen: !this.state.advancedOpen})}>Advanced
-              Options</a>
+               onClick={(e) => {
+                 e.preventDefault();
+                 this.setState({advancedOpen: !this.state.advancedOpen})
+               }}
+            >
+              Advanced Options
+            </a>
             <InputWrapper
               style={{display: this.state.advancedOpen ? 'block' : 'none'}}>
               <InputWrapper>
@@ -132,7 +135,6 @@ export class ViewFieldWidget extends Component {
                   helperText={`Separate contextual filters with a "/". Each filter may use "+" or "," for multi-value arguments. This field supports tokens.`}
                   variant="outlined"
                   defaultValue={this.state['arguments']}
-                  required={this.props.settings.required}
                   onChange={e => this.valueChanged('arguments', e.target.value)}
                   inputProps={{maxLength: 254}}
                   fullWidth
@@ -154,7 +156,7 @@ export class ViewFieldWidget extends Component {
               </InputWrapper>
             </InputWrapper>
           </AdvancedOptions>
-        </InputWrapper>
+        </InputGroupWrapper>
       </FormGroup>
     )
   }
@@ -164,7 +166,14 @@ export class ViewFieldWidget extends Component {
 const AdvancedOptions = styled.div`
   border: 1px solid #ccc;
   padding: 20px;
+  margin-top: 20px;
 `;
+
+const InputGroupWrapper = styled.div`
+  border: 1px solid #ccc;
+  padding: 20px;
+  margin-top: 10px;
+`
 
 const InputWrapper = styled.div`
   margin-top: 10px;