+++ src/Plugin/Field/FieldWidget/ReactParagraphs.php
@@ -7,6 +7,8 @@ use Drupal\Component\Utility\NestedArray;
 use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\field\FieldConfigInterface;
+use Drupal\paragraphs\ParagraphInterface;
+use Drupal\react_paragraphs\Entity\ParagraphsRowInterface;
 
 /**
  * Plugin implementation of the 'react_paragraphs' widget.
@@ -65,7 +67,7 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
   /**
    * {@inheritdoc}
    */
-  public function settingsSummary() {
+  public function settingsSummary(): array {
     return [
       $this->getSetting('resizable') ? $this->t('Resizable') : $this->t('Equal Widths'),
     ];
@@ -74,7 +76,7 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
   /**
    * {@inheritdoc}
    */
-  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
+  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state): array {
     $element['container'] = $element;
     $element['container'] += [
       '#type' => 'fieldset',
@@ -135,7 +137,7 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    * @return array
    *   Keyed data of rows and items.
    */
-  protected function getRowItems(FieldItemListInterface $items) {
+  protected function getRowItems(FieldItemListInterface $items): array {
     $row_item_field = self::getRowItemsField($this->fieldDefinition);
     $all_items = [];
 
@@ -173,12 +175,11 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
   /**
    * {@inheritdoc}
    */
-  public function massageFormValues(array $values, array $form, FormStateInterface $form_state) {
+  public function massageFormValues(array $values, array $form, FormStateInterface $form_state): array {
     if (!empty($this->rowData)) {
       return $this->rowData;
     }
-
-    $react_data = json_decode(urldecode($values['container']['value']), TRUE);
+    $react_data = json_decode(rawurldecode($values['container']['value']), TRUE);
 
     $return_data = [];
 
@@ -232,14 +233,15 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    * @param int|null $entity_id
    *   Entity id if its existing.
    *
-   * @return \Drupal\Core\Entity\ContentEntityInterface
+   * @return \Drupal\react_paragraphs\Entity\ParagraphsRowInterface
    *   Row entity with set field values.
    *
    * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  protected function getRowEntity(array $field_data, $entity_id) {
+  protected function getRowEntity(array $field_data, ?int $entity_id): ParagraphsRowInterface {
+    /** @var \Drupal\react_paragraphs\Entity\ParagraphsRowInterface $entity */
     $entity = $this->getEntity('paragraph_row', $this->getRowBundle(), $field_data, $entity_id);
     $entity->save();
     return $entity;
@@ -250,20 +252,17 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    *
    * @param array $field_data
    *   Array of field data from the react side.
-   * @param int $width
+   * @param int|null $width
    *   Width of the current item.
-   * @param string $admin_label
+   * @param string|null $admin_label
    *   Administrative label.
    * @param int|null $entity_id
    *   Entity id.
    *
    * @return \Drupal\paragraphs\ParagraphInterface
    *   Entity with modified field values.
-   *
-   * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  protected function getRowItemEntity(array $field_data, $width, $admin_label, $entity_id) {
-    /** @var \Drupal\paragraphs\ParagraphInterface $row_item */
+  protected function getRowItemEntity(array $field_data, ?int $width, ?string $admin_label, ?int $entity_id): ParagraphInterface {
     $row_item = $this->getEntity('paragraph', $field_data['type'][0]['target_id'], $field_data, $entity_id);
     $row_item->setBehaviorSettings('react', [
       'width' => $width,
@@ -279,7 +278,7 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    * @return string
    *   Bundle ID.
    */
-  protected function getRowBundle() {
+  protected function getRowBundle(): string {
     $settings = $this->fieldDefinition->getSetting('handler_settings');
     return reset($settings['target_bundles']);
   }
@@ -296,18 +295,18 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    * @param int|null $entity_id
    *   Entity ID if available.
    *
-   * @return \Drupal\Core\Entity\ContentEntityInterface
+   * @return \Drupal\paragraphs\ParagraphInterface
    *   Entity with modified field values.
    *
    * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    */
-  protected function getEntity($entity_type, $bundle, array $field_data, $entity_id = NULL) {
+  protected function getEntity(string $entity_type, string $bundle, array $field_data, ?int $entity_id = NULL): ParagraphInterface {
     if (!$entity_id) {
       return $this->createEntity($entity_type, $bundle, $field_data);
     }
 
-    /** @var \Drupal\Core\Entity\ContentEntityInterface $entity */
+    /** @var \Drupal\paragraphs\ParagraphInterface $entity */
     $entity = $this->entityTypeManager->getStorage($entity_type)
       ->load($entity_id);
     $entity->setNewRevision();
@@ -318,6 +317,8 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
         $entity->set($field_name, $field_data[$field_name]);
       }
     }
+
+    $this->setBehaviorSettings($entity, $field_data['behavior_settings'] ?? []);
     return $entity;
   }
 
@@ -331,13 +332,13 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
    * @param array $field_data
    *   Array of field data from the react side.
    *
-   * @return \Drupal\Core\Entity\ContentEntityInterface
+   * @return \Drupal\paragraphs\ParagraphInterface
    *   Newly created entity.
    *
    * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    */
-  protected function createEntity($entity_type, $bundle, array $field_data) {
+  protected function createEntity(string $entity_type, string $bundle, array $field_data): ParagraphInterface {
     $entity_definition = $this->entityTypeManager->getDefinition($entity_type);
     $bundle_key = $entity_definition->getKey('bundle');
     $storage = $this->entityTypeManager->getStorage($entity_type);
@@ -349,12 +350,29 @@ class ReactParagraphs extends ReactParagraphsWidgetBase {
         $entity_values[$field_name] = $field_data[$field_name];
       }
     }
-    /** @var \Drupal\Core\Entity\ContentEntityInterface $entity */
+    /** @var \Drupal\paragraphs\ParagraphInterface $entity */
     $entity = $storage->create($entity_values);
+    $this->setBehaviorSettings($entity, $field_data['behavior_settings'] ?? []);
 
     // TODO: Handle validation.
     $entity->validate();
     return $entity;
+  }
+
+  /**
+   * Set the entity's behavior settings based on the values from the UI.
+   *
+   * @param \Drupal\paragraphs\ParagraphInterface $entity
+   *   Created or edited paragraph/row entity.
+   * @param array $behaviors
+   *   Keyed array of behavior values.
+   */
+  protected function setBehaviorSettings(ParagraphInterface $entity, array $behaviors) {
+    if (!empty($behaviors[0]['value'])) {
+      foreach ($behaviors[0]['value'] as $plugin_id => $settings) {
+        $entity->setBehaviorSettings($plugin_id, $settings);
+      }
+    }
   }
 
 }